Factorial of a Number
The "Factorial of a Number" program calculates the factorial of a given integer using a recursive function. The factorial of a non-negative integer 
�
n is the product of all positive integers less than or equal to 
�
n. The program consists of two main parts: the factorial function and the main function.